#39 file login-modal style added
