+++ app/components/login-modal/index.jsx
@@ -4,69 +4,7 @@ var bootstrap = require("react-bootstrap");
 
 var components = require("components");
 var colors     = require("lib/colors");
-
-var styles = {
-    center: {
-        textAlign: "center"
-    },
-    overlay: {
-        position: "fixed",
-        top: "0px",
-        left: "0px",
-        height: "100%",
-        width: "100%",
-        background: colors.primary,
-        zIndex: 1000
-    },
-    h4Label: {
-        marginBottom: "100px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    accessButton: {
-        fontWeight: "bold",
-        height: "50px",
-        color: colors.white,
-        fontSize: "20px"
-    },
-    titleLabel: {
-        marginTop: "200px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    inputLabel: {
-        height: "50px",
-        fontSize: "18px"
-        // opacity: 0.1
-    },
-    h4PasswordLostLabel: {
-        marginBottom: "30px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    groupIcon: {
-        width: "36px"
-    },
-    inputsContainer: {
-        position: "absolute",
-        left: "calc(50% - 175px)",
-        width: "350px"
-    },
-    textTitlePosition: {
-        width: "350px"
-    },
-    popupLabel: {
-        color: colors.white
-    },
-    aLink: {
-        color: colors.white,
-        fontSize: "15px"
-    },
-    errorAlert: {
-        marginTop: "16px",
-        textAlign: "center"
-    }
-};
+var loginStyle = require("lib/login-modal-style")
 
 var LoginModal = React.createClass({
     propTypes: {
@@ -141,7 +79,7 @@ var LoginModal = React.createClass({
         return this.state.mailError ? (
             <bootstrap.Alert
                 bsStyle="danger"
-                style={styles.errorAlert}
+                style={loginStyle.errorAlert}
             >
                 {"Hai sbagliato a inserire la mail!"}
             </bootstrap.Alert>
@@ -151,7 +89,7 @@ var LoginModal = React.createClass({
         return this.state.loginError ? (
             <bootstrap.Alert
                 bsStyle="danger"
-                style={styles.errorAlert}
+                style={loginStyle.errorAlert}
             >
                 {"Hai sbagliato a inserire mail o password!"}
             </bootstrap.Alert>
