+++ app/components/login-modal/login-page.jsx
@@ -4,69 +4,7 @@ var bootstrap = require("react-bootstrap");
 
 var components = require("components");
 var colors     = require("lib/colors");
-
-var styles = {
-    center: {
-        textAlign: "center"
-    },
-    overlay: {
-        position: "fixed",
-        top: "0px",
-        left: "0px",
-        height: "100%",
-        width: "100%",
-        background: colors.primary,
-        zIndex: 1000
-    },
-    h4Label: {
-        marginBottom: "100px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    accessButton: {
-        fontWeight: "bold",
-        height: "50px",
-        color: colors.white,
-        fontSize: "20px"
-    },
-    titleLabel: {
-        marginTop: "200px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    inputLabel: {
-        height: "50px",
-        fontSize: "18px"
-        // opacity: 0.1
-    },
-    h4PasswordLostLabel: {
-        marginBottom: "30px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    groupIcon: {
-        width: "36px"
-    },
-    inputsContainer: {
-        position: "absolute",
-        left: "calc(50% - 175px)",
-        width: "350px"
-    },
-    textTitlePosition: {
-        width: "350px"
-    },
-    popupLabel: {
-        color: colors.white
-    },
-    aLink: {
-        color: colors.white,
-        fontSize: "15px",
-        cursor: "pointer"
-    },
-    errorAlert: {
-        marginTop: "16px"
-    }
-};
+var loginStyle = require("lib/login-modal-style");
 
 var LoginPage = React.createClass({
     propTypes: {
@@ -76,19 +14,19 @@ var LoginPage = React.createClass({
     },
     render: function () {
         return (
-            <div style={styles.overlay}>
+            <div style={loginStyle.overlay}>
                 <components.StyleLogin />
-                <span className="text-center" style={styles.textTitlePosition}>
-                    <h1 style={styles.titleLabel}>{"Energia alla tua Energia"}</h1>
-                    <h4 style={styles.h4Label}>{"Innowatio"}</h4>
+                <span className="text-center" style={loginStyle.textTitlePosition}>
+                    <h1 style={loginStyle.titleLabel}>{"Energia alla tua Energia"}</h1>
+                    <h4 style={loginStyle.h4Label}>{"Innowatio"}</h4>
                 </span>
-                <div style={styles.inputsContainer}>
+                <div style={loginStyle.inputsContainer}>
                     <bootstrap.Input
-                        addonBefore={<components.Icon icon="user" style={styles.groupIcon}/>}
+                        addonBefore={<components.Icon icon="user" style={loginStyle.groupIcon}/>}
                         className="form-signin-email"
                         placeholder="Email"
                         ref="email"
-                        style={styles.inputLabel}
+                        style={loginStyle.inputLabel}
                         type="email"
                         required>
                     </bootstrap.Input>
@@ -96,13 +34,13 @@ var LoginPage = React.createClass({
                         <bootstrap.Input
                             addonBefore={
                                 <span className="iconPsw">
-                                    <components.Icon className="iconPsw" icon="lock" style={styles.groupIcon}/>
+                                    <components.Icon className="iconPsw" icon="lock" style={loginStyle.groupIcon}/>
                                 </span>
                                 }
                             className="form-signin-psw"
                             placeholder="Password"
                             ref="password"
-                            style={styles.inputLabel}
+                            style={loginStyle.inputLabel}
                             type="password"
                         />
                     </div>
@@ -110,14 +48,14 @@ var LoginPage = React.createClass({
                         block
                         className="access-button"
                         onClick={this.props.onChange}
-                        style={styles.accessButton}
+                        style={loginStyle.accessButton}
                     >
                         {"ACCEDI"}
                     </bootstrap.Button>
                     {this.props.errorMail}
                     <components.Spacer direction="v" size={10} />
-                    <div className="text-center" style={styles.popupLabel}>
-                        <a onClick={this.props.lostPassword} style={styles.aLink}>{"Username o Password dimenticati?"}</a>
+                    <div className="text-center" style={loginStyle.popupLabel}>
+                        <a onClick={this.props.lostPassword} style={loginStyle.aLink}>{"Username o Password dimenticati?"}</a>
                     </div>
                 </div>
             </div>
