+++ app/components/login-modal/request-password-page.jsx
@@ -4,68 +4,7 @@ var bootstrap = require("react-bootstrap");
 
 var components = require("components");
 var colors     = require("lib/colors");
-
-var styles = {
-    center: {
-        textAlign: "center"
-    },
-    overlay: {
-        position: "fixed",
-        top: "0px",
-        left: "0px",
-        height: "100%",
-        width: "100%",
-        background: colors.primary,
-        zIndex: 1000
-    },
-    h4Label: {
-        marginBottom: "100px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    accessButton: {
-        fontWeight: "bold",
-        height: "50px",
-        color: colors.white,
-        fontSize: "20px"
-    },
-    titleLabel: {
-        marginTop: "200px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    inputLabel: {
-        height: "50px",
-        fontSize: "18px"
-        // opacity: 0.1
-    },
-    h4PasswordLostLabel: {
-        marginBottom: "30px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    groupIcon: {
-        width: "36px"
-    },
-    inputsContainer: {
-        position: "absolute",
-        left: "calc(50% - 175px)",
-        width: "350px"
-    },
-    textTitlePosition: {
-        width: "350px"
-    },
-    popupLabel: {
-        color: colors.white
-    },
-    aLink: {
-        color: colors.white,
-        fontSize: "15px"
-    },
-    errorAlert: {
-        marginTop: "16px"
-    }
-};
+var loginStyle = require("lib/login-modal-style");
 
 var RequestPasswordPage = React.createClass({
     propTypes: {
@@ -74,35 +13,34 @@ var RequestPasswordPage = React.createClass({
     },
     render: function () {
         return (
-            <div style={styles.overlay}>
+            <div style={loginStyle.overlay}>
                 <components.StyleLogin />
-                <span className="text-center" style={styles.textTitlePosition}>
-                    <h1 style={styles.titleLabel}>{"Energia alla tua Energia"}</h1>
-                    <h4 style={styles.h4PasswordLostLabel}>{"Innowatio"}</h4>
-                    <h4 style={styles.h4PasswordLostLabel}>{"Username o Password dimenticati?"}</h4>
+                <span className="text-center" style={loginStyle.textTitlePosition}>
+                    <h1 style={loginStyle.titleLabel}>{"Energia alla tua Energia"}</h1>
+                    <h4 style={loginStyle.h4PasswordLostLabel}>{"Innowatio"}</h4>
+                    <h4 style={loginStyle.h4PasswordLostLabel}>{"Username o Password dimenticati?"}</h4>
                 </span>
-                <div style={styles.inputsContainer}>
+                <div style={loginStyle.inputsContainer}>
                     <bootstrap.Input
-                        addonBefore={<components.Icon icon="user" style={styles.groupIcon}/>}
+                        addonBefore={<components.Icon icon="user" style={loginStyle.groupIcon}/>}
                         className="form-signin-email"
                         placeholder="Email"
                         ref="email"
-                        style={styles.inputLabel}
+                        style={loginStyle.inputLabel}
                         type="email"
                         required>
                     </bootstrap.Input>
-                    {console.log(this.props.ref)}
                     <bootstrap.Button
                         block
                         className="access-button"
                         onClick={this.props.onChange}
-                        style={styles.accessButton}
+                        style={loginStyle.accessButton}
                     >
                         {"INVIA"}
                     </bootstrap.Button>
                     {this.props.errorMail}
-                    <span className="text-center" style={styles.textTitlePosition}>
-                        <h6 style={styles.h4PasswordLostLabel}>{"Riceverai le nuove credenziali all'indirizzo e-mail indicato"}</h6>
+                    <span className="text-center" style={loginStyle.textTitlePosition}>
+                        <h6 style={loginStyle.h4PasswordLostLabel}>{"Riceverai le nuove credenziali all'indirizzo e-mail indicato"}</h6>
                     </span>
                 </div>
             </div>
