+++ app/components/login-modal/sent-password-page.jsx
@@ -4,55 +4,7 @@ var bootstrap = require("react-bootstrap");
 
 var components = require("components");
 var colors     = require("lib/colors");
-
-var styles = {
-    center: {
-        textAlign: "center"
-    },
-    overlay: {
-        position: "fixed",
-        top: "0px",
-        left: "0px",
-        height: "100%",
-        width: "100%",
-        background: colors.primary,
-        zIndex: 1000
-    },
-    h4Label: {
-        marginBottom: "100px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    accessButton: {
-        fontWeight: "bold",
-        height: "50px",
-        color: colors.white,
-        fontSize: "20px"
-    },
-    titleLabel: {
-        marginTop: "200px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    inputLabel: {
-        height: "50px",
-        fontSize: "18px"
-        // opacity: 0.1
-    },
-    h4PasswordLostLabel: {
-        marginBottom: "30px",
-        color: colors.white,
-        fontWeight: "100"
-    },
-    inputsContainer: {
-        position: "absolute",
-        left: "calc(50% - 175px)",
-        width: "350px"
-    },
-    textTitlePosition: {
-        width: "350px"
-    }
-};
+var loginStyle = require("lib/login-modal-style");
 
 var SentPasswordPage = React.createClass({
     propTypes: {
@@ -60,21 +12,21 @@ var SentPasswordPage = React.createClass({
     },
     render: function () {
         return (
-            <div style={styles.overlay}>
+            <div style={loginStyle.overlay}>
                 <components.StyleLogin />
-                <span className="text-center" style={styles.textTitlePosition}>
-                    <h1 style={styles.titleLabel}>{"Energia alla tua Energia"}</h1>
-                    <h4 style={styles.h4PasswordLostLabel}>{"Innowatio"}</h4>
-                    <h4 style={styles.h4PasswordLostLabel}>
+                <span className="text-center" style={loginStyle.textTitlePosition}>
+                    <h1 style={loginStyle.titleLabel}>{"Energia alla tua Energia"}</h1>
+                    <h4 style={loginStyle.h4PasswordLostLabel}>{"Innowatio"}</h4>
+                    <h4 style={loginStyle.h4PasswordLostLabel}>
                         {"Abbiamo mandato una mail all'indirizzo selezionato per il reset della password."}
                     </h4>
                 </span>
-                <div style={styles.inputsContainer}>
+                <div style={loginStyle.inputsContainer}>
                     <bootstrap.Button
                         block
                         className="access-button"
                         onClick={this.props.onChange}
-                        style={styles.accessButton}
+                        style={loginStyle.accessButton}
                     >
                         {"TORNA ALLA LOGIN"}
                     </bootstrap.Button>
