+++ app/lib/login-modal-style.js
@@ -1,0 +1,68 @@
+var colors   = require("lib/colors");
+var bkgImage = require("lib/image/image!background.jpg")
+
+var styles = {
+    center: {
+        textAlign: "center"
+    },
+    overlay: {
+        position: "fixed",
+        top: "0px",
+        left: "0px",
+        height: "100%",
+        width: "100%",
+        backgroundImage: ,
+        zIndex: 1000
+    },
+    h4Label: {
+        marginBottom: "100px",
+        color: colors.white,
+        fontWeight: "100"
+    },
+    accessButton: {
+        fontWeight: "bold",
+        height: "50px",
+        color: colors.white,
+        fontSize: "20px"
+    },
+    titleLabel: {
+        marginTop: "200px",
+        color: colors.white,
+        fontWeight: "100"
+    },
+    inputLabel: {
+        height: "50px",
+        fontSize: "18px"
+        // opacity: 0.1
+    },
+    h4PasswordLostLabel: {
+        marginBottom: "30px",
+        color: colors.white,
+        fontWeight: "100"
+    },
+    groupIcon: {
+        width: "36px"
+    },
+    inputsContainer: {
+        position: "absolute",
+        left: "calc(50% - 175px)",
+        width: "350px"
+    },
+    textTitlePosition: {
+        width: "350px"
+    },
+    popupLabel: {
+        color: colors.white
+    },
+    aLink: {
+        color: colors.white,
+        fontSize: "15px",
+        cursor: "pointer"
+    },
+    errorAlert: {
+        marginTop: "16px",
+        textAlign: "center"
+    }
+};
+
+module.exports = styles;
